Add message if project is not installed

+++ Makefile
@@ -15,7 +15,7 @@ clean:
 	rm -fr dist
 
 clean-install: clean
-	rm -fr node_modules
+	rm -fr $(NODE_MODULES)
 	rm -fr $(VENV)
 
 lint-python:
@@ -51,4 +51,5 @@ build-check:
 	set -m; ((STATIC_SERVER= $(FLASK) run -h $(HOST) -p $(PYTHON_PORT); kill 0)& ($(NPM) run render-server; kill 0)& wait)
 
 serve:
+	test -d $(NODE_MODULES) || (echo 'Please run make install before serving.' && exit 1)
 	set -m; (($(FLASK) run -h $(HOST) -p $(PYTHON_PORT); kill 0)& ($(NPM) run static-server; kill 0)& ($(NPM) run render-server; kill 0)& wait)
